+++ options/options.js
@@ -231,11 +231,9 @@ function generateBehaviorConfigs() {
         'backgroundDimmerOpacity',
         'storeCurrentScrollPosition',
         'highlightElementOnHover',
-
         'continiousVerticalScrollDetection',
         'continiousHorizontalScrollDetection',
-
-        'animateHideRelativeToSelected',
+        // 'animateHideRelativeToSelected',
     ];
 
     inputIds.forEach(function (inputId) {
@@ -1162,8 +1160,7 @@ function updateDisabledOptions() {
     document.querySelector("#delayToShowTitleOnHoverWhenHidden").parentNode.className = document.querySelector("#showTitleOnHoverWhenHidden").checked ? 'visible-option' : 'hidden-option';
     document.querySelector("#backgroundDimmerOpacity").parentNode.className = document.querySelector("#dimBackground").checked ? 'visible-option' : 'hidden-option';
     document.querySelector("#hideCircleIfNoActionSelected").parentNode.className = document.getElementById('openCircleOn').value == 'rightClick' ? 'enabled-option' : 'disabled-option';
-
-    document.querySelector("#animateHideRelativeToSelected").parentNode.className = document.querySelector("#circleHideAnimation").checked ? 'visible-option' : 'hidden-option';
+    // document.querySelector("#animateHideRelativeToSelected").parentNode.className = document.querySelector("#circleHideAnimation").checked ? 'visible-option' : 'hidden-option';
 
 }