+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Added separate configs for icons and labels opacity
 - Added 'Circle location' config (always under cursor / always corner / under cursor + in corner when no space / under cursor + shift when no space)
 - Resolved conflict with drag when 'long left click to open' is active
+- Fixed issue when circle still had hide animation despite it's disabled in settings
 - Changed some default configs
 - Small options page improvements